- added services fighting fix -> still experimental code - revisited DoShutdown() code - added putlog() with reason - did some misc changes around

+++ include/chanserv.h
@@ -179,7 +179,7 @@ void SetDefaultALVL(struct ChanInfo *cptr);
 
 extern struct ChanInfo *chanlist[CHANLIST_MAX];
 extern struct Channel *ChannelList;
-extern int MaxTSDelta;
+extern long MaxTSDelta;
 
 #endif /* CHANNELSERVICES */
